test: reorganize test logic
- more concise helpers
- acknowledge failing tests (temporarily)
- delete config after each test
